--- checks/e2e/citeria/empty/Cargo.toml
@@ -1,16 +1,0 @@
-[package]
-name = "migration-e2e-test-empty-criterion"
-version = { workspace = true }
-edition = { workspace = true }
-license = { workspace = true }
-authors = { workspace = true }
-homepage = { workspace = true }
-publish = { workspace = true }
-rust-version = { workspace = true }
-
-[dependencies]
-migration-e2e-test-types = { workspace = true }
-
-
-[lints]
-workspace = true
