--- checks/e2e/citeria/empty/README.md
@@ -1,2 +1,0 @@
-# Empty
-The empty criterion asserts nothing about the migration. It is the free criterion.
